Use mktemp as output, add (Zoomed) text for [url] mode, small changes
- mktemp allows not modifying the last edited timestamp on the directory
as well as parallel runs of spectro on the same directory.

- Add '(Zoomed)' text for the [url] title to avoid both links having
the same title (filename).

- Change the "Uploading $filename to imgur" text to remove the filename
part, which is not very informative when it comes from mktemp. It now
uses the source's filename and the (Zoomed) text, when appropriate.

- Fix issue where the SHA would be shown twice when using --double, -d.

- Fix issue with --double, -d in offline mode (--local, -l)  where runs
after the first file would use the '_zoomed' filename (thus causing
overwrites).

- Add and edit some comments for clarity.