Remove duplicated lines for offline/local mode
The part of the script where it showed the URL was written twice in the
script: Once for local mode and the other one for online. Changing the
if to a negative one and moving those lines to the end of the function
removes the redundancy.
This commit also fixes a couple of missed variable name changes from
last commit (from $f to $file).