Use basename instead of full argument
This avoids cases where the full path to the file would be displayed as
the title of the spectrogram or as the name for the output image when
using --local. Now only the name for the file will be used, and for the
directory shown in the title, only the directory directly above the file
will be shown (e.g. before it could show full/path/to/DirectoryWithMusic
and now it will only show DirectoryWithMusic/file.flac)