Allow AAC, WMA, ape, mp4, add quiet option and more
Accept WMA, .ape and MP4 files (uses ffmpeg or avconv, whichever finds
first). Using the converter it will first create a WAV from the input in
a temporal file, which will then be processed as spectro normally
processes WAV files. After the spectrogram is created the .WAV file that
was generated is removed.

Add -q, --quiet flag to remove the info about the upload process (this
allows redirection with | or > keeping the same format)

Use hash instead of command for portability when checking for
dependencies.

Add converter info to SoX's comment (e.g. 'Created with spectro using
SoX and ffmpeg')