Add --compare function, modularise existing code
The --compare flag will allow any number of files as input. From each
file it will create a spectrogram which will be added to a .gif file at
the end. This .gif is the only output displayed (the other spectrograms
are deleted once the .gif is created). This allows for easy comparison
between different encoders, for example. The gif is created using
ImageMagick's 'convert' utility, which will be credited on the SoX image
comment.

Already existing code was moved to the functions upload_to_imgur and
show_output, as to not repeat lines of code for the --compare function.

Error messages start with 'ERROR: ' now, for consistency.